+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - L2CAP ERTM: assert that packet can be stored in fragment buffers in ertm
+- L2CAP ERTM: reject incoming connection if ERTM is configured as mandatory but not supported by remote
 
 ### Added
 - GAP: allow to reject GAP classic connection via gap_register_classic_connection_filter